Rename Dry::View::Controller to Dry::View
