+++ examples/rails/app/views/application_view.rb
@@ -1,4 +1,4 @@
-class ApplicationView < Dry::View::Controller
+class ApplicationView < Dry::View
   config.paths = Rails.root.join("app/templates")
   config.default_context = ApplicationViewContext.new
 end
