+++ lib/dry/view/tilt/haml.rb
@@ -1,5 +1,5 @@
 module Dry
-  module View
+  class View
     module Tilt
       module Haml
         def self.requirements
