+++ lib/dry/view/version.rb
@@ -1,5 +1,5 @@
 module Dry
-  module View
+  class View
     VERSION = '0.5.4'.freeze
   end
 end
