+++ spec/fixtures/integration/scopes/custom_view_scope.html.slim
