+++ spec/fixtures/templates/view_renderer_options.html.erb
