+++ spec/integration/scope_spec.rb
@@ -2,13 +2,13 @@ require "dry/view/part"
 require "dry/view/scope"
 
 RSpec.describe "Scopes" do
-  let(:base_vc) {
-    Class.new(Dry::View::Controller) do
+  let(:base_view) {
+    Class.new(Dry::View) do
       config.paths = FIXTURES_PATH.join("integration/scopes")
     end
   }
 
-  specify "Custom scope for a view controller" do
+  specify "Custom scope for a view" do
     module Test
       class ControllerScope < Dry::View::Scope
         def hello
@@ -17,24 +17,24 @@ RSpec.describe "Scopes" do
       end
     end
 
-    vc = Class.new(base_vc) do
-      config.template = "custom_view_controller_scope"
+    view = Class.new(base_view) do
+      config.template = "custom_view_scope"
       config.scope = Test::ControllerScope
 
       expose :text
     end.new
 
-    expect(vc.(text: "world").to_s).to eq "Hello world!"
+    expect(view.(text: "world").to_s).to eq "Hello world!"
   end
 
   specify "Rendering a partial via an anonymous scope" do
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.template = "anonymous_scope"
 
       expose :text
     end.new
 
-    expect(vc.(text: "Hello").to_s).to eq "Greeting: Hello"
+    expect(view.(text: "Hello").to_s).to eq "Greeting: Hello"
   end
 
   specify "Rendering a partial implicitly via a custom named scope" do
@@ -46,14 +46,14 @@ RSpec.describe "Scopes" do
       end
     end
 
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.scope_namespace = Test::Scopes
       config.template = "named_scope_with_implicit_render"
 
       expose :text
     end.new
 
-    expect(vc.(text: "Hello").to_s).to eq "Greeting: HELLO!"
+    expect(view.(text: "Hello").to_s).to eq "Greeting: HELLO!"
   end
 
   specify "Rendering a partial implicitly via a custom named scope (provided via a class)" do
@@ -65,22 +65,22 @@ RSpec.describe "Scopes" do
       end
     end
 
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.scope_namespace = Test::Scopes
       config.template = "class_named_scope_with_implicit_render"
 
       expose :text
     end.new
 
-    expect(vc.(text: "Hello").to_s).to eq "Greeting: HELLO!"
+    expect(view.(text: "Hello").to_s).to eq "Greeting: HELLO!"
   end
 
   specify "Raising an error when an unnamed partial cannot be rendered implicitly" do
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.template = "unnamed_named_scope_with_implicit_render"
     end.new
 
-    expect { vc.().to_s }.to raise_error ArgumentError, "+partial_name+ must be provided for unnamed scopes"
+    expect { view.().to_s }.to raise_error ArgumentError, "+partial_name+ must be provided for unnamed scopes"
   end
 
   specify "Rendering a partial explicitly via a custom named scope" do
@@ -92,14 +92,14 @@ RSpec.describe "Scopes" do
       end
     end
 
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.scope_namespace = Test::Scopes
       config.template = "named_scope_with_explicit_render"
 
       expose :text
     end.new
 
-    expect(vc.(text: "Hello").to_s).to eq "Holler: HELLO!"
+    expect(view.(text: "Hello").to_s).to eq "Holler: HELLO!"
   end
 
   specify "Custom named scope providing defaults for missing locals" do
@@ -111,14 +111,14 @@ RSpec.describe "Scopes" do
       end
     end
 
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.scope_namespace = Test::Scopes
       config.template = "named_scope_with_defaults"
 
       expose :text
     end.new
 
-    expect(vc.().to_s).to eq "Greeting: Howdy"
+    expect(view.().to_s).to eq "Greeting: Howdy"
   end
 
   specify "Creating a custom scope from a view part" do
@@ -138,7 +138,7 @@ RSpec.describe "Scopes" do
       end
     end
 
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.part_namespace = Test::Parts
       config.scope_namespace = Test::Scopes
       config.template = "scope_from_part"
@@ -146,6 +146,6 @@ RSpec.describe "Scopes" do
       expose :message
     end.new
 
-    expect(vc.(message: {text: "Hello from a part"}).to_s).to eq "Greeting: Hello from a part!"
+    expect(view.(message: {text: "Hello from a part"}).to_s).to eq "Greeting: Hello from a part!"
   end
 end
