+++ spec/integration/template_engines/erubi_spec.rb
@@ -2,12 +2,12 @@ require "erubi"
 require "erubi/capture_end"
 require "tilt/erubi"
 
+require "dry/view"
 require "dry/view/context"
-require "dry/view/controller"
 
 RSpec.describe "Template engines / erb (using erubi via an explict engine mapping)" do
-  let(:base_vc) {
-    Class.new(Dry::View::Controller) do
+  let(:base_view) {
+    Class.new(Dry::View) do
       config.paths = FIXTURES_PATH.join("integration/template_engines/erubi")
       config.renderer_engine_mapping = {erb: Tilt::ErubiTemplate}
       config.renderer_options = {engine_class: Erubi::CaptureEndEngine}
@@ -15,11 +15,11 @@ RSpec.describe "Template engines / erb (using erubi via an explict engine mappin
   }
 
   it "supports partials that yield" do
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.template = "render_and_yield"
     end.new
 
-    expect(vc.().to_s.gsub(/\n\s*/m, "")).to eq "<wrapper>Yielded</wrapper>"
+    expect(view.().to_s.gsub(/\n\s*/m, "")).to eq "<wrapper>Yielded</wrapper>"
   end
 
   it "supports context methods that yield" do
@@ -29,11 +29,11 @@ RSpec.describe "Template engines / erb (using erubi via an explict engine mappin
       end
     end.new
 
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.default_context = context
       config.template = "method_with_yield"
     end.new
 
-    expect(vc.().to_s.gsub(/\n\s*/m, "")).to eq "<wrapper>Yielded</wrapper>"
+    expect(view.().to_s.gsub(/\n\s*/m, "")).to eq "<wrapper>Yielded</wrapper>"
   end
 end
