+++ spec/integration/template_engines/slim_spec.rb
@@ -1,20 +1,20 @@
 require "slim"
+require "dry/view"
 require "dry/view/context"
-require "dry/view/controller"
 
 RSpec.describe "Template engines / slim" do
-  let(:base_vc) {
-    Class.new(Dry::View::Controller) do
+  let(:base_view) {
+    Class.new(Dry::View) do
       config.paths = FIXTURES_PATH.join("integration/template_engines/slim")
     end
   }
 
   it "supports partials that yield" do
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.template = "render_and_yield"
     end.new
 
-    expect(vc.().to_s).to eq "<wrapper>Yielded</wrapper>"
+    expect(view.().to_s).to eq "<wrapper>Yielded</wrapper>"
   end
 
   it "supports context methods that yield" do
@@ -24,11 +24,11 @@ RSpec.describe "Template engines / slim" do
       end
     end.new
 
-    vc = Class.new(base_vc) do
+    view = Class.new(base_view) do
       config.default_context = context
       config.template = "method_with_yield"
     end.new
 
-    expect(vc.().to_s).to eq "<wrapper>Yielded</wrapper>"
+    expect(view.().to_s).to eq "<wrapper>Yielded</wrapper>"
   end
 end
